+++ CHANGELOG.md
@@ -11,6 +11,15 @@ Gotta blame the broken mods!
       
     Dispenser Behaviors:
 •  Reduced the logspam from dispenser registry replacement detection.
+  
+    Missing Loot Table:
+•  Removed logspam when jeresources checks other mod's entities that doesn't have loottables.
+
+•  Added 1 extra stacktrace line to missing loottable reports in case that might help more
+   on finding where it is called.
+     
+    DynamicRegistries:
+• Removed stacktrace from Classloading DynamicRegistry if it is done by Vampirism as they do it safely.
 
 
    (V.1.7.3 Changes) (1.16.2/1.16.3/1.16.4 Minecraft)